Change ContainerState.exitCode from Integer to Long
updated CHANGELOG and bumped minor version.

see #1052

+++ src/main/java/com/spotify/docker/client/messages/ContainerState.java
@@ -56,7 +56,7 @@ public abstract class ContainerState {
   public abstract Integer pid();
 
   @JsonProperty("ExitCode")
-  public abstract Integer exitCode();
+  public abstract Long exitCode();
 
   @JsonProperty("StartedAt")
   public abstract Date startedAt();
@@ -83,7 +83,7 @@ public abstract class ContainerState {
       @JsonProperty("Paused") final Boolean addr,
       @JsonProperty("Restarting") final Boolean restarting,
       @JsonProperty("Pid") final Integer pid,
-      @JsonProperty("ExitCode") final Integer exitCode,
+      @JsonProperty("ExitCode") final Long exitCode,
       @JsonProperty("StartedAt") final Date startedAt,
       @JsonProperty("FinishedAt") final Date finishedAt,
       @JsonProperty("Error") final String error,
